+++ examples/bridge/src/types/mod.rs
@@ -1,0 +1,3 @@
+pub mod block;
+pub mod inbound;
+pub mod outbound;
